Add support for using postMessage API
Setting the URL to https://slack.com/api/chat.postMessage endpoint and
providing a valid token will post a message using this method instead
of a webhook. This method returns the channel ID and timestamp, which
are both needed to post a subsequent message as a reply in a thread.

Note that channel_file has been removed and channel accepts a single
channel ID, whereas before it would loop through a space-separated
list of IDs.

TODO add docs for token parameter.